refactor(groups): extract GroupHitlCoordinator from GroupConversationService
R1 step 7: unite the pause/cancel/timeout helpers and the cancelDiscussion/
resumeDiscussion public surface into one collaborator. Widens executeDiscussion,
resolvePhases and cleanupEphemeralAgents to public so the coordinator can call
back into the facade, same self-reference pattern as MemberTurnExecutor.

GroupConversationService: 2,594 -> 1,885 lines. 7 of R1's 10 steps done.
